+++ app/src/main/java/com/example/librasheet/viewModel/dataClasses/CategoryUi.kt
@@ -6,7 +6,6 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.graphics.Color
 import com.example.librasheet.data.entity.Category
 import com.example.librasheet.data.entity.CategoryId
-import com.example.librasheet.data.toFloatDollar
 import com.example.librasheet.ui.graphing.PieChartValue
 
 
@@ -16,22 +15,22 @@ data class CategoryUi(
     val key: Long = 0,
     val id: CategoryId = CategoryId(),
     override val color: Color = Color.White,
-    override val value: Float = 0f,
+    override val value: Double = 0.0,
     val subCategories: List<CategoryUi> = emptyList(),
 ) : PieChartValue {
     override val name: String
         get() = id.name
 }
 
-fun Category.toUi(values: Map<Long, Float>, multiplier: Float = 1f): CategoryUi {
+fun Category.toUi(values: Map<Long, Double>, multiplier: Float = 1f): CategoryUi {
     val subs = subCategories.map { it.toUi(values, multiplier) }
-    val value = multiplier * values.getOrDefault(key, 0f) + subs.sumOf { it.value.toDouble() }.toFloat()
+    val value = multiplier * values.getOrDefault(key, 0.0) + subs.sumOf { it.value }
     return CategoryUi(
         category = this,
         key = key,
         id = id,
         color = color,
-        value = if (value == 0f) 0f else value, // this is needed so that we don't have -$0.00
+        value = if (value == 0.0) 0.0 else value, // this is needed so that we don't have -$0.00
         subCategories = subs
     )
 }
